+++ odbc-api/src/guide.rs
@@ -50,10 +50,10 @@
 //!
 //! Oh dear! Aren't we of to a bad start. First step in using this API and already a piece of unsafe
 //! code. Well we are talking with a C API those contract explicitly demands that there may be only
-//! one ODBC Environment in the entire process. The only way to know that this is call is safe is
-//! in application code. If you write a library you MUST NOT wrap the creation of an ODBC
-//! environment in a safe function call. If another libray would do the same and an application were
-//! to use both of these, it might create two environments in safe code and thus causing undefined
+//! one ODBC Environment in the entire process. The only way to know that this call is safe is in
+//! application code. If you write a library you MUST NOT wrap the creation of an ODBC environment
+//! in a safe function call. If another libray would do the same and an application were to use
+//! both of these, it might create two environments in safe code and thus causing undefined
 //! behaviour, which is clearly a violation of Rusts safety guarantees.
 //!
 //! Apart from that. This is it. Our ODBC Environment is ready for action.